share executeRedisFsApiWithMetric
